refactor(cmd): name regex package after its directory
